+++ src/poke/mod.rs
@@ -100,14 +100,10 @@ pub fn load_session_settings(
         "bns".to_string(),
     ];
     settings.initial_deployer = initial_deployer;
-    settings.costs_version = project_config.project.costs_version;
+    settings.repl_settings = project_config.repl_settings.clone();
     settings.disk_cache_enabled = true;
 
-    if let Some(ref analysis_passes) = project_config.project.analysis {
-        settings.analysis = analysis_passes.clone();
-    }
-
-    settings.parser_version = project_config.project.parser_version;
+    settings.repl_settings.parser_version = settings.repl_settings.parser_version;
 
     Ok((settings, chain_config, project_config))
 }