+++ src/types/project_manifest.rs
@@ -1,4 +1,5 @@
 use crate::utils;
+use clarity_repl::repl;
 use std::collections::{BTreeMap, HashSet};
 use std::env::var;
 use std::fs::File;
@@ -12,6 +13,7 @@ use toml::value::Value;
 pub struct ProjectManifestFile {
     project: ProjectConfigFile,
     contracts: Option<Value>,
+    repl: Option<repl::SettingsFile>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -21,7 +23,10 @@ pub struct ProjectConfigFile {
     description: Option<String>,
     telemetry: Option<bool>,
     requirements: Option<Value>,
-    analysis: Option<Vec<String>>,
+
+    // The fields below have been moved into repl above, but are kept here for
+    // backwards compatibility.
+    analysis: Option<Vec<clarity_repl::analysis::Pass>>,
     costs_version: Option<u32>,
     parser_version: Option<u32>,
     cache_dir: Option<String>,
@@ -32,6 +37,8 @@ pub struct ProjectManifest {
     pub project: ProjectConfig,
     #[serde(serialize_with = "toml::ser::tables_last")]
     pub contracts: BTreeMap<String, ContractConfig>,
+    #[serde(rename = "repl")]
+    pub repl_settings: repl::Settings,
 }
 
 #[derive(Serialize, Deserialize, Debug, Default)]
@@ -41,9 +48,6 @@ pub struct ProjectConfig {
     pub description: String,
     pub telemetry: bool,
     pub requirements: Option<Vec<RequirementConfig>>,
-    pub analysis: Option<Vec<String>>,
-    pub costs_version: u32,
-    pub parser_version: u32,
     pub cache_dir: String,
 }
 
@@ -119,18 +123,38 @@ impl ProjectManifest {
                 .unwrap_or("".into()),
             authors: project_manifest_file.project.authors.unwrap_or(vec![]),
             telemetry: project_manifest_file.project.telemetry.unwrap_or(false),
-            costs_version: project_manifest_file.project.costs_version.unwrap_or(2),
-            analysis: project_manifest_file.project.analysis,
-            parser_version: project_manifest_file.project.parser_version.unwrap_or(2),
             cache_dir: project_manifest_file
                 .project
                 .cache_dir
                 .unwrap_or(default_cache_path),
         };
 
+        let mut repl_settings = if let Some(repl_settings) = project_manifest_file.repl {
+            repl::Settings::from(repl_settings)
+        } else {
+            repl::Settings::default()
+        };
+
+        // Check for deprecated settings
+        if let Some(passes) = project_manifest_file.project.analysis {
+            println!(
+                "{}: use of 'project.analysis' in Clarinet.toml is deprecated; use repl.analysis.passes",
+                yellow!("warning")
+            );
+            repl_settings.analysis.set_passes(passes);
+        }
+        if let Some(costs_version) = project_manifest_file.project.costs_version {
+            println!(
+                "{}: use of 'project.costs_version' in Clarinet.toml is deprecated; use repl.costs_version",
+                yellow!("warning")
+            );
+            repl_settings.costs_version = costs_version;
+        }
+
         let mut config = ProjectManifest {
             project,
             contracts: BTreeMap::new(),
+            repl_settings,
         };
         let mut config_contracts = BTreeMap::new();
         let mut config_requirements: Vec<RequirementConfig> = Vec::new();